Refactor 2D array allocation and deallocation functions
Change the parameter type of rows and cols from int to size_t in allocate_2d_array and destroy_2d_array functions

Update the function signature and implementation accordingly

+++ src/brick_game/tetris/modules/init.c
@@ -11,7 +11,7 @@
 
 #include "../include/init.h"
 
-void allocate_2d_array(int **array, int rows, int cols) {
+void allocate_2d_array(int **array, size_t rows, size_t cols) {
   array = (int **)calloc(rows, sizeof(int *));
 
   if (!array) {
@@ -19,7 +19,7 @@ void allocate_2d_array(int **array, int rows, int cols) {
     MEM_ALLOC_ERROR
   }
 
-  for (int i = 0; i < rows; i++) {
+  for (size_t i = 0; i < rows; i++) {
     array[i] = (int *)calloc(cols, sizeof(int));
     if (!array[i]) {
       destroy_2d_array(array, rows, cols);
@@ -28,9 +28,9 @@ void allocate_2d_array(int **array, int rows, int cols) {
   }
 }
 
-void destroy_2d_array(int **array, int rows, int cols) {
+void destroy_2d_array(int **array, size_t rows, size_t cols) {
   if (array) {
-    for (int i = 0; i < rows; i++) {
+    for (size_t i = 0; i < rows; i++) {
       if (array[i]) {
         free(array[i]);
         array[i] = NULL;
